# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Barberia",
        "description": "Servicio de API tipo SOA para implementación en una barberia",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/ticket/nuevo": {
            "post": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Crear un nuevo ticket",
                "operationId": "c7cc94024ea5760acf1a37aa564f011d",
                "requestBody": {
                    "description": "Datos del ticket para crear",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "IdDetallesGeneral": {
                                        "description": "ID del detalle general",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "fecha": {
                                        "description": "Fecha del ticket",
                                        "type": "string",
                                        "example": "2023-02-23"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Se ha creado el ticket correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Los datos a validar no son correctos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Los datos a validar no son correctos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al crear el ticket",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al crear el ticket."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "api/ticket/idTicket/editar": {
            "put": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Actualizar un ticket específico",
                "operationId": "a09c70c8d1af0b317495abb2189e2635",
                "parameters": [
                    {
                        "name": "idTicket",
                        "in": "path",
                        "description": "ID del ticket a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del ticket para actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "IdDetallesGeneral": {
                                        "description": "ID del detalle general",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "fecha": {
                                        "description": "Fecha del ticket",
                                        "type": "string",
                                        "example": "2023-02-23"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El ticket ha sido actualizado correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Los datos a validar no son correctos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Los datos a validar no son correctos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al actualizar el ticket",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al actualizar el ticket."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/ticket/idTicket": {
            "get": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Mostrar un ticket específico",
                "operationId": "791ba3d6d83b2c9c9b8c14156e0f60b4",
                "parameters": [
                    {
                        "name": "idTicket",
                        "in": "path",
                        "description": "ID del ticket a mostrar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Idticket": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "IdDetallesGeneral": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "fecha": {
                                            "type": "string",
                                            "example": "2023-02-23"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al mostrar el ticket",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al mostrar el ticket."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ticket/idTicket/eliminar": {
            "delete": {
                "tags": [
                    "Ticket"
                ],
                "summary": "Eliminar un ticket específico",
                "operationId": "971652fb9e363d4e2480b21ff8ef26ab",
                "parameters": [
                    {
                        "name": "idTicket",
                        "in": "path",
                        "description": "ID del ticket a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El ticket ha sido eliminado correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ha ocurrido un error al eliminar el ticket",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ha ocurrido un error al eliminar el ticket."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}